requirements.txt
python-telegram-bot==13.15
pandas==2.2.0
urllib3==1.26.15
requests
certifi
